MP2-2218 adjust height of section C.6 to support multipage table export

See merge request interact/ems/jems-default-plugin!69

+++ src/main/kotlin/io/cloudflight/jems/plugin/standard/application_form_export/timeplan/TranslatedText.kt
@@ -2,7 +2,7 @@ package io.cloudflight.jems.plugin.standard.application_form_export.timeplan
 
 data class TranslatedText(
     val x: Float,
-    val y: Int,
+    val y: Float,
     val value: String,
     val dominantBaseline: String = "middle",
     val fontStyle: String = "normal",
